+++ Plugin.cs
@@ -20,6 +20,7 @@ namespace LootManager
         internal static HashSet<string> Whitelist = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         internal static HashSet<string> Banklist  = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         internal static HashSet<string> Junklist  = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+        internal static HashSet<string> Auctionlist = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         internal static HashSet<string> Editlist = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         
         public static Dictionary<string, HashSet<string>> FilterList =
@@ -82,6 +83,7 @@ namespace LootManager
             LootWhitelist.Load();
             LootBanklist.Load();
             LootJunklist.Load();
+            LootAuctionlist.Load();
             LootFilterlist.Load();
 
             Log.LogInfo("Loot Manager loaded.");